Clean up syntax, spelling, types

+++ src/Msg.cpp
@@ -6,7 +6,7 @@
 
 #include "Msg.h"
 
-CMsg::CMsg(char cType, const char * pMsg, DWORD dwTime)
+CMsg::CMsg(char cType, const char * pMsg, uint64_t dwTime)
 {
 	m_cType = cType;
     message = pMsg;
@@ -14,7 +14,7 @@ CMsg::CMsg(char cType, const char * pMsg, DWORD dwTime)
 	m_iObjectID = -1;
 }
 
-CMsg::CMsg(char cType, std::string_view msg, DWORD dwTime)
+CMsg::CMsg(char cType, std::string_view msg, uint64_t dwTime)
 {
     m_cType = cType;
     message = msg;
@@ -22,7 +22,7 @@ CMsg::CMsg(char cType, std::string_view msg, DWORD dwTime)
     m_iObjectID = -1;
 }
 
-CMsg::CMsg(char cType, std::string & msg, DWORD dwTime)
+CMsg::CMsg(char cType, std::string & msg, uint64_t dwTime)
 {
     m_cType = cType;
     message = msg;
